Rename: owcurves.py -> owspectra.py, OWCurves -> OWSpectra
