+++ orangecontrib/infrared/widgets/owintegrate.py
@@ -17,7 +17,7 @@ from Orange.widgets import gui, settings
 from orangecontrib.infrared.data import getx
 from orangecontrib.infrared.preprocess import Integrate
 
-from orangecontrib.infrared.widgets.owcurves import SELECTONE
+from orangecontrib.infrared.widgets.owspectra import SELECTONE
 from orangecontrib.infrared.widgets.owhyper import refresh_integral_markings
 from orangecontrib.infrared.widgets.owpreproc import SetXDoubleSpinBox, MovableVlineWD
 import orangecontrib.infrared.widgets.owpreproc
